apps/document-generator/pyproject.toml
[project]
name = "document-generator"
version = "0.1.0"
description = "Gradio app for using the document-generator recipe"
authors = [{ name = "MADE:Explorations Team" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.30.0",
    "jsonschema>=4.0.0",
    "msal>=1.32.3",
    "pydantic>=1.10.0",
    "python-dotenv>=0.20.0",
    "recipe-tool>=0.1.1",
    "typer>=0.7.0",
]

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "debugpy>=1.8.14",
    "pyright>=1.1.400",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.10",
]

[project.scripts]
document-generator-app = "document_generator_app.main:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["document_generator_app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
